Added Special 2 Area
Special 2 Area is a riddle.
After killing DatDolan, he instantiates a portal with particle effects that leads to the Special 2 Area. There, there is always one special red MainEnemy that drops either a 0 or a 1 when dead. The Player has to solve the riddle by putting the numbers on the right spots. He only has one chance to do so. He gets rewarded with 2000 Kleines Yen, if he succeeds.